update blog controller tests
